NY 브랜치 따서 커밋 & 머지리퀘스트요청 - widgets 폴더 만들고 밑에다가 widgets들 분리해서 넣음
